correção das infrações

+++ biblioteca/UserEligibilityHandler.java
@@ -3,7 +3,7 @@ package biblioteca;
 public class UserEligibilityHandler extends AprovationHandler {
     @Override
     public boolean handleRequest(User user, Copy copy) {
-        int infractions = user.getInfractions();
+        double infractions = user.getInfractions();
         if (infractions > 2) {
             System.out.println("Número de infrações excedido!");
             return false;
